🚧 get converted texture data and display as image
