+++ package.json
@@ -13,7 +13,8 @@
     "three": "^0.140.0"
   },
   "devDependencies": {
-    "@react-spring/three": "^9.4.4",
+    "@react-spring/core": "^9.4.5",
+    "@react-spring/three": "^9.4.5",
     "@react-three/drei": "^9.6.0",
     "@react-three/fiber": "^8.0.12",
     "@types/react": "^18.0.0",
