+++ src/components/CubicCameras.tsx
@@ -5,14 +5,13 @@ import * as THREE from 'three'
 
 type CubicCamerasProps = {
 	cameraList: MutableRefObject<THREE.PerspectiveCamera>[]
-	helpers?: boolean
 }
-const CubicCameras = ({ cameraList, helpers = false }: CubicCamerasProps) => {
+const CubicCameras = ({ cameraList }: CubicCamerasProps) => {
 	useEffect(() => {
 		cameraList.forEach(camera => {
 			camera.current.aspect = 1
 			camera.current.updateProjectionMatrix()
-		})
+		}) // run everytime this component re-render
 	})
 
 	useEffect(() => {
@@ -24,7 +23,7 @@ const CubicCameras = ({ cameraList, helpers = false }: CubicCamerasProps) => {
 
 		cameraList[4].current.lookAt(0, 0, 1) // pz
 		cameraList[5].current.lookAt(0, 0, -1) // nz
-	}, [cameraList])
+	}, [])
 
 	const { rotation } = useControls({
 		output: folder({
@@ -37,6 +36,20 @@ const CubicCameras = ({ cameraList, helpers = false }: CubicCamerasProps) => {
 		}),
 	})
 
+	const { cameraHelpers } = useControls({
+		helpers: folder(
+			{
+				cameraHelpers: {
+					value: false,
+					label: 'camera helpers',
+				},
+			},
+			{
+				collapsed: true,
+			},
+		),
+	})
+
 	return (
 		<group rotation-y={rotation}>
 			{cameraList.map((_, index) => (
@@ -46,7 +59,7 @@ const CubicCameras = ({ cameraList, helpers = false }: CubicCamerasProps) => {
 					args={[90, 1, 0.1, 100]}
 				></PerspectiveCamera>
 			))}
-			{helpers &&
+			{cameraHelpers &&
 				cameraList.map((camera, index) => <cameraHelper key={index} args={[camera.current]} />)}
 		</group>
 	)
