+++ src/components/StatefulAxesHelper.tsx
@@ -1,0 +1,14 @@
+import { folder, useControls } from 'leva'
+
+export function StatefulAxesHelper() {
+	const { axesHelpers } = useControls({
+		helpers: folder({
+			axesHelpers: {
+				value: false,
+				label: 'axes helper',
+			},
+		}),
+	})
+
+	return <axesHelper visible={axesHelpers}></axesHelper>
+}
